% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_guide_lines.R
\name{add_guide_lines}
\alias{add_guide_lines}
\title{Add labeled guide-lines to assist cowplot item position}
\usage{
add_guide_lines(p, axis = "x", min = 0, max = 1, n = 5, signif = 3)
}
\arguments{
\item{p}{ggplot object. Object to add plot lines to.}

\item{min, max}{[0,1]. Minimum and maximum values for lines, with 0 being the left-hand side/bottom
and 1 being the right-hand side/bottom.}

\item{n}{positive integer. Number of lines to draw. Default is 5.}

\item{"x", }{"y" or "xy". Axi(e)s lines should be perpendicular to.}
}
\value{
Object of class "gg". However, it cannot be treated as a standard
"gg" object, and so any manipulations (adding geoms, changing axis labels, etc.)
will not work on this plot.
}
\description{
When using `cowplot::ggdraw` and associated functions, occasionally
one needs to be very precise about where to place items. However, it can be difficult
to eyeball the exact coordinates, and re-running the entire plot multiple times can be
slow. This function draws lines that have their x- or y-dimensions labeled,
allowing one to gradually zoom in in on the desired coordinates very accurately and quickly.
}
\examples{

library(ggplot2)
p <- ggplot(data.frame(x = 1, y = 1)) + geom_point()
ggutils::add_cp_grid_guide(p = p, signif = 2)

}
