% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_p_value.R
\name{add_p_value}
\alias{add_p_value}
\title{Add p-values to a plot}
\usage{
add_p_value(
  p,
  p_txt,
  q_txt = NULL,
  line_x,
  p_nudge,
  q_nudge,
  y,
  tick_nudge_in = 0.01,
  tick_height = 0.01,
  text_size = 14,
  line_size = 1
)
}
\arguments{
\item{p}{ggplot object. Object to add p-values to.}

\item{p_txt, q_txt}{character. P-value and q-value text,
e.g. "p = 0.01" and "(q = 0.05)".}

\item{line_x}{c([0,1], [0,1]). Specify left and right endpoints of line along
the x-axis.}

\item{p_nudge, q_nudge}{c([0,1]). By default, the p and q-values texts are
centred at the middle of \code{line_x}. They may be shifted horizontally using
these parameters, with positive values shifting them further to the right.}

\item{y}{c([0,1], [0,1], [0,1]). Specifies vertical coordinates of the
p-value, q-value and the line. Note that all three are required, even
if the q-value is not used.}

\item{tick_nudge_in}{[0,1]. Amount by which to push the line's ticks inwards.
Default is 0.01. Will often need to be adjusted slightly.}

\item{tick_height}{[0,1]. Height of ticks. Default is 0.01.}

\item{text_size, line_size}{positive numeric. Size of text and line, respectively.
Defaults ar 14 and 1, respctively.}
}
\value{
Object of class "gg". However, it cannot be treated as a standard
"gg" object, and so any manipulations (adding geoms, changing axis labels, etc.)
will not work on this object. They should therefore be done before, with the
plot afterward consisted as fixed apart from further additions from `cowplot::draw_`
functions.
}
\description{
Add p-values to a plot
}
