% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcorr.R
\name{ggcorr}
\alias{ggcorr}
\title{Plot scatterplot with correlation coefficients and ab-line overlaid}
\usage{
ggcorr(
  data,
  corr_method = "spearman",
  corr_lab = "auto",
  grp,
  grp_base = NULL,
  grp_x = NULL,
  grp_lab = "auto",
  y,
  id,
  est_signif = 3,
  ci = TRUE,
  pval = TRUE,
  ci_signif = 3,
  pval_signif = 3,
  pval_trunc = 0.001,
  label_id = FALSE,
  label_id_size = 3,
  trans = "identity",
  coord = c(0.05, 0.95),
  skip = 0.05,
  font_size = 4.2,
  hjust = 0,
  vjust = 0.5,
  point_size = 1,
  point_alpha = 0.9,
  limits_expand = NULL,
  limits_equal = FALSE,
  legend_title = FALSE,
  legend_position = "bottom",
  thm = cowplot::theme_cowplot() + theme(plot.background = element_rect(fill = "white")),
  grid = cowplot::background_grid(major = "xy"),
  grp_to_col = NULL,
  abline = FALSE,
  smooth = TRUE,
  smooth_method = "lm",
  smooth_se = FALSE
)
}
\arguments{
\item{data}{dataframe. Dataframe containing the data to be plotted.}

\item{corr_method}{character vector. Correlation method. One or more of
\code{c("concordance", "pearson", "spearman", "kendall")}. Default is \code{"spearman"}.}

\item{corr_lab}{\verb{"never"``, }"auto"``, \code{"always"} or a named character vector.
Whether, and how, to label the correlation methods in the plot.
If "always" or "never", then the correlation labels are always or never
displayed, respectively.
If `"auto"`, then they are displayed only if multiple correlation methods
are used.
If displaye dusing `"always"` or `"never"`, then default labels are used.
If a named character vector, then the names are the correlation methods and
the values are the labels,
e.g. `c("spearman" = "Spearman", "kendall" = "Kendall")`.
Default is `"auto"`.
correlation methods are used.}

\item{grp}{character. Name of the column in \code{data} that specifies the
variables to compare between. The first level within \code{grp} (i.e. \code{data[[grp]][1]})
is the level plotted on the x-axis, unless \code{grp_x} is set (see below).}

\item{grp_base}{character. If not \code{NULL}, then only correlations between this
level within \code{data[[grp]]} and other levels are displayed. If \code{NULL}, all
two-way correlations are displayed. Default is \code{NULL}.}

\item{grp_x}{character. If \code{NULL}, then this level of \code{grp} is plotted along
the x-axis. Default is \code{NULL}.}

\item{grp_lab}{\code{"always"}, \code{"auto"} or \code{"never"}. Whether to label the groups
being compared for a specific displayed correlation coefficients.
If \code{"always"} or \code{"never"}, then the groups are always or never labelled,
respectively.
If \code{"auto"}, then the groups are labelled only if multiple groups
are compared.
Default is \code{"auto"}.}

\item{y}{character. Name of the column in \code{data} that specifies the
values being correlated between groups.}

\item{id}{character. Name of the column in \code{data} that specifies the subject
from which multiple measurements were taken.}

\item{est_signif}{numeric. Significant digits for correlation estimates.
Default is \code{3}.}

\item{ci}{logical. Whether to include confidence intervals for correlation
estimates. Default is \code{TRUE}.}

\item{pval}{logical. Whether to include p-values for correlation estimates.
Default is \code{TRUE}.}

\item{ci_signif}{numeric. Significant digits for confidence intervals.
Default is \code{3}.}

\item{pval_signif}{numeric. Significant digits for p-values. Default is \code{3}.}

\item{pval_trunc}{numeric. P-values smaller than this threshold will be
displayed as "< threshold". Default is \code{0.001}.}

\item{label_id}{logical. If \code{TRUE}, points labels as given by \code{data[[id]]}
are displayed using \code{ggrepel::geom_text_repel}. Default is \code{FALSE}.}

\item{label_id_size}{numeric. Size of labels. Default is \code{3}.}

\item{trans}{character or trans object. Specifies scaling of y- and x-axes.
If class is character, it is converted to a trans object. Adds "root_cube",
"root_fourth", "root_fifth" and "asinh" transformations, as well as "sqrt"
transformation that allows plotting of lines at zero. If class is a trans
object, it is returned as is. Default is \code{"identity"}.}

\item{coord}{numeric vector.
Coordinates for text placement for first (or only) row of results.
Specified in terms of proportion of axis length,
with the first value specifying the x-axis coordinate and the second
value the y-axis coordinate.
Default is \code{c(0.05, 0.95)}, which places the first row
in the top left corner of the plot.}

\item{skip}{numeric.
Proportion of y-axis to move down between lines of text.
Default is \code{0.05}.}

\item{font_size}{numeric. Font size for text in results table. Default is \code{10}.}

\item{hjust}{numeric. Horizontal justification for text. Default is \code{0}.}

\item{vjust}{numeric. Vertical justification for text. Default is \code{0.5}.}

\item{point_size}{numeric. Size of points. Default is \code{1}.}

\item{point_alpha}{numeric. Alpha value for points. Default is \code{0.9}.}

\item{limits_expand}{list. Expand the axis limits. Default is \code{NULL}.}

\item{limits_equal}{logical. Set axis limits equal. Default is \code{FALSE}.}

\item{legend_title}{logical.
If \code{FALSE}, then the legend title is removed. Default is \code{FALSE}.}

\item{legend_position}{character. Position of the legend. Default is \code{"bottom"}.
If \code{NULL}, then the legend's positioning is not jadjusted.}

\item{thm}{ggplot2 theme. Specifies \code{ggplot2} theme. Default is
\code{cowplot::theme_cowplot()},
with a white plot background (as opposed to transparent).}

\item{grid}{The value from calling \code{cowplot::background_grid}.
Specifies display of grid lines.
If \code{NULL}, then no background grid is displayed
(recommended for non-cowplot-themed plots).
Default is \code{cowplot::background_grid(major = "xy")}, which displays
major grid lines on both x- and y-axes.}

\item{grp_to_col}{character vector.
Specifies colours for elements in \code{grp}.
Elements are colours.
If named, the names specify the level within \code{grp} to
which the colours are matched.
If \code{NULL} and only two groups are compared,
then all points are black.
If \code{NULL} and there are more than two groups
compared, then defaults to palette \code{Set1} of the RColorBrewer package.
Default is \code{NULL}.}

\item{abline}{logical. If \code{TRUE}, then the y=x line is plotted in colour
"gray85". Default is \code{TRUE}.}

\item{smooth}{logical. If \code{TRUE}, then a linear regression line is plotted.}

\item{smooth_method}{character. Smoothing method to pass to \code{geom_smooth}.
Default is \code{"lm"}.}

\item{smooth_se}{logical. Whether to display 95\% confidence interval for the
smooth line. Default is \code{FALSE}.}
}
\description{
Plots a scatterplot with various correlation coefficients and an optional
ab-line overlaid.
}
\examples{
response_vec_a <- rnorm(5)
response_tbl <- data.frame(
  group = rep(letters[1:3], each = 5),
  response = c(
    response_vec_a,
    response_vec_a * 1.2 + rnorm(5, sd = 0.2),
    response_vec_a * 2 + rnorm(5, sd = 2)
  ),
  pid = rep(paste0("id_", 1:5), 3)
)
library(UtilsGGSV)
ggcorr(
  data = response_tbl,
  grp = "group",
  y = "response",
  id = "pid"
)
}
