% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_text_column.R
\name{get_trans}
\alias{get_trans}
\title{Get trans object from character}
\usage{
get_trans(trans)
}
\arguments{
\item{trans}{character or trans object.
If class is character, then it is converted to a trans object.
Adds "root_cube", "root_fourth", "root_fifth" and
"asinh" transformations, as well as "sqrt" transformation
that allows plotting of lines at zero.
If class is a trans object, then it is returned as is.}
}
\description{
Get trans object from character
}
\examples{
x_vec <- seq(1, 5, length.out = 1e3)
y_vec <- UtilsGGSV::get_trans("root_fifth")$transform(seq_vec)
plot_tbl <- data.frame(x = x_vec, y = y_vec)
ggplot(
  plot_tbl,
  aes(x, y)
) +
  geom_line() +
  geom_point() +
  coord_equal() +
  expand_limits(x = 5, y = 5)
}
